Add support for responses generated in the future
ChatProcessor becomes a protocol that only contains the `help` of a
processor. You additionally conform to FutureResponder or
SynchronousResponder, with SynchronousResponder corresponding to the
existing -> ChatResponse? interface. FutureResponders instead return
FutureChatResponse?.

FutureChatResponse, in turn, acts as a primitive, unsynchronized promise
for a ChatResponse. It is fulfilled externally, and that triggers a
response. ReplBot implements these two alternatives via a switch
statement, and relevant helpers are strapped in throughout.