Add GordiasBrain module/library
GordiasBrain provides the basics of a brain for the operations system.
Brain is currently a very simple protocol that allows reading an Any
value from and writing an Any value to a String subscript. For now, we
provide an InMemoryBrain that is just a typealias for a conforming
Dictionary<String,Any>.