Replace mapRange/forEachRange/etc with ranges(inString:)
The block iterators provided custom ways to iterate and map over
substring matches of a given string; however, the much simpler and more
general way of doing this is to provide a way to access to all of the
ranges in an existing sequence.

ranges(inString:) builds on range(inString:at:) to return an array of
all matches as Substrings of the String it is passed. This allows all
collection methods to be applied normally to the returned Array.