SavingBrains are saved every second by ReplBot
Clean/dirty tracking is up to the brain if it wants to do it; if a
SavingBrain is passed to ReplBot, then a timer fires every second that
calls save() on the SavingBrain.

Currently, a failed save (by throwing) will not back off on save
attempts. Additionally, the timer isn't cleaned up in the bot's deinit,
though it's worth noting that the bot is expected to last the lifetime
of the process currently.